requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
Werkzeug==2.0.1
plotly==5.1.0
pandas==1.3.0
python-dotenv==0.19.0
email-validator==1.1.3
Jinja2==3.0.1
SQLAlchemy==1.4.23
WTForms==2.3.3
